+++ packages/data-access/src/index.ts
@@ -2,4 +2,3 @@ export { default as DataAccess } from './data-access';
 export { default as TransactionIndex } from './transaction-index';
 export { default as Block } from './block';
 export { CombinedDataAccess } from './combined-data-access';
-export { ReadonlyDataAccess } from './readonly-data-access';
