+++ packages/request-node/src/thegraph-node.ts
@@ -7,7 +7,7 @@ import { RequestNodeBase } from './requestNodeBase';
 import * as config from './config';
 import { getIpfsStorage } from './storageUtils';
 import Utils from '@requestnetwork/utils';
-import { TheGraphDataAccess } from '@requestnetwork/thegraph-storage';
+import { TheGraphDataAccess, TheGraphStorage } from '@requestnetwork/thegraph-storage';
 
 export class TheGraphRequestNode extends RequestNodeBase {
   constructor(url: string, logger?: LogTypes.ILogger) {
@@ -27,12 +27,17 @@ export class TheGraphRequestNode extends RequestNodeBase {
     );
     const signer = new NonceManager(wallet);
     const ipfsStorage = getIpfsStorage(logger);
+    const storage = new TheGraphStorage({
+      ipfsStorage,
+      signer,
+      network,
+      logger,
+    });
     const dataAccess = new TheGraphDataAccess({
       graphql: { url },
-      ipfsStorage,
+      storage,
       network,
-      signer,
-      logger: logger,
+      logger,
     });
 
     super(dataAccess, ipfsStorage, store, logger);
