+++ packages/thegraph-storage/src/index.ts
@@ -3,9 +3,8 @@ export {
   TheGraphDataRead,
   TheGraphDataWrite,
   TheGraphDataAccessOptions,
-  TheGraphReadonlyDataAccess,
-} from './TheGraphDataAccess';
-export { PendingStore } from './PendingStore';
-export { TheGraphStorage } from './TheGraphStorage';
-export { SubgraphClient } from './subgraphClient';
+} from './data-access';
+export { PendingStore } from './pending-store';
+export { TheGraphStorage } from './storage';
+export { SubgraphClient } from './subgraph-client';
 export * as queries from './queries';
